requirements.txt
asgiref==3.4.1
click==8.0.3
falcon==3.0.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
psycopg2-binary==2.9.2
python-dotenv==0.19.2
PyYAML==6.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
